env args and done
