+++ Makefile
@@ -6,7 +6,7 @@
 #    By: asyed <asyed@student.42singapore.sg>       +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/05/24 16:11:35 by ting              #+#    #+#              #
-#    Updated: 2024/06/04 18:10:08 by asyed            ###   ########.fr        #
+#    Updated: 2024/06/06 16:55:08 by asyed            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -22,7 +22,7 @@ SRCSPATH = srcs/
 
 SRCS = minishell.c builtins_env_node_utils.c builtin_env.c\
 		builtins_utils.c builtins.c del_and_free.c\
-		builtin_export.c builtin_unset.c
+		builtin_export.c builtin_unset.c builtin_exit.c
 #SRCS = minishell.c parsing.c \
 		lexer.c expand_env_var.c \
 		remove_quotes.c free_for_lexer.c \
